\name{optidefix-package}
\alias{optidefix-package}
\alias{optidefix}
\docType{package}
\title{
\packageTitle{optidefix}
}
\description{
\packageDescription{optidefix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{optidefix}
\packageIndices{optidefix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{optidefix}

Maintainer: \packageMaintainer{optidefix}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
